Fixed framerate independence Timer now uses Chrono header
Fixes #5

+++ src/bounce/Object.h
@@ -64,7 +64,7 @@ public:
 	PlayerObject(IFade2D* lib, vec2 Len, vec2 Pos, std::string texturePath, float angle = 0.f);
 	void move(vec2 mov);
 	void moveTo(vec2 pos);
-	void update(vec2 accel, double timeStep, std::vector<Object> worldObjects);
+	void update(vec2 accel, float timeStep, std::vector<Object> worldObjects);
 	BoundingCircle boundingCircle;
 private:
 	float radius;
